+++ src/main.cpp
@@ -36,7 +36,7 @@ int main(int argc, char* argv[]) {
 
     /* Process file content */
     int matrix_size = reader.get_data_size();
-    std::vector<std::vector<int>> dist(matrix_size, std::vector<int>(matrix_size, 0));
+    std::vector<std::vector<float>> dist(matrix_size, std::vector<float>(matrix_size, 0.0));
     FilePreprocessor file_preprocessor;
     file_preprocessor.create_dist_matrix(content, matrix_size, dist);
